Final Commit

+++ src/main/java/com/fujitsu/service/UserService.java
@@ -1,7 +1,12 @@
 package com.fujitsu.service;
 
+import com.fujitsu.config.BlockListConfig;
+import com.fujitsu.exception.InvalidEmailExceptions;
 import com.fujitsu.model.User;
 import com.fujitsu.repository.UserRepository;
+import org.apache.commons.validator.routines.EmailValidator;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,11 +15,31 @@ import java.util.List;
 @Service
 public class UserService {
 
+    private static Logger LOGGER = LoggerFactory.getLogger(UserService.class);
     @Autowired
     private UserRepository userRepository;
 
-    public User createUser(User user){
-        return userRepository.save(user);
+    @Autowired
+    private BlockListConfig blocklist;
+
+    public User createUser(User user) throws InvalidEmailExceptions {
+        LOGGER.debug("Blocklist size"+blocklist.getBlockList().size());
+
+        if(!EmailValidator.getInstance().isValid(user.getEmailAddress())){
+            throw new InvalidEmailExceptions("Invalid email address:"+user.getEmailAddress());
+        }
+
+        String domain = user.getEmailAddress().substring(user.getEmailAddress().indexOf("@") + 1);
+        if(blocklist.getBlockList().contains(domain)){
+           throw new InvalidEmailExceptions("Email Address is in block list. Cannot create user for this email address:"+user.getEmailAddress());
+        }
+
+        if(isEmailAlreadyExist(user)){
+            throw new InvalidEmailExceptions("User already exist. Cannot create user for this email address:"+user.getEmailAddress());
+        }else{
+            return userRepository.save(user);
+        }
+
     }
 
     public User getUser(User user){
@@ -26,11 +51,15 @@ public class UserService {
     }
 
     public String deleteUser(User user){
-        userRepository.delete(user);
-        return "User deleted";
+        if(userRepository.existsById(user.getId())){
+            userRepository.delete(user);
+            return "User deleted";
+        }else{
+            return "User do not exist, nothing to delete";
+        }
     }
 
     public boolean isEmailAlreadyExist(User user){
-        userRepository.findByEmailAddress(user.getEmailAddress()).
+        return userRepository.existsUserByEmailAddress(user.getEmailAddress());
     }
 }
